+++ Sources/MapboxNavigation/NavigationMapView+LocationAuthorization.swift
@@ -1,0 +1,6 @@
+extension NavigationMapView: LocationAuthorization {
+    
+    func isAuthorized() -> Bool {
+        return allowedAuthorizationStatuses.contains(authorizationStatus)
+    }
+}